+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -116,8 +116,9 @@ Changes in existing checks
   <clang-tidy/checks/readability/redundant-smartptr-get>` check to
   remove `->`, when redundant `get()` is removed.
 
-- Added option `UseUpperCaseLiteralSuffix` to :doc:`readablility-implicit-bool-conversion
-  <clang-tidy/checks/readability/implicit-bool-conversion>` check to select the
+- Improved :doc:`readablility-implicit-bool-conversion
+  <clang-tidy/checks/readability/implicit-bool-conversion>` check
+  Added option `UseUpperCaseLiteralSuffix` to  to select the
   case of the literal suffix in fixes.
 
 Removed checks